+++ src/f10_world/test/test_world_etl0_jungle_to_zoo.py
@@ -1,33 +1,18 @@
 from src.f00_instrument.file import create_path
-from src.f04_gift.atom_config import (
-    face_id_str,
-    fiscal_id_str,
-    acct_id_str,
-    owner_id_str,
-)
+from src.f04_gift.atom_config import face_id_str, fiscal_id_str
 from src.f07_fiscal.fiscal_config import (
     cumlative_minute_str,
     hour_label_str,
     weekday_label_str,
     weekday_order_str,
 )
-from src.f08_pidgin.pidgin_config import (
-    event_id_str,
-    inx_wall_str,
-    otx_wall_str,
-    inx_acct_id_str,
-    otx_acct_id_str,
-    unknown_word_str,
-    inx_label_str,
-    otx_label_str,
-)
+from src.f08_pidgin.pidgin_config import event_id_str
 from src.f09_brick.pandas_tool import (
     _get_pidgen_brick_format_filenames,
     get_sheet_names,
     upsert_sheet,
 )
-from src.f10_world.world import worldunit_shop, _create_events_agg_df
-from src.f10_world.world_tool import get_all_brick_dataframes
+from src.f10_world.world import worldunit_shop
 from src.f10_world.examples.world_env import get_test_worlds_dir, env_dir_setup_cleanup
 from pandas import DataFrame, read_excel as pandas_read_excel
 from os.path import exists as os_path_exists
